Hw5: minor scrapper refactoring
